Get color scheme from thecolorapi.com API

Return a generated scheme for the provided seed color and optional mode. Color schemes are multi-color combinations chosen according to color-wheel relationsships.

Usage

get_color_thecolorapi_scheme(
  hex = NULL,
  rgb = NULL,
  hsl = NULL,
  cmyk = NULL,
  count = 5,
  mode = "monochrome",
  format = "json",
  w = 100,
  named = TRUE,
  return_name_only = TRUE
)

Arguments

hex

string (optional), Valid hex code.

rgb

string (optional), Valid rgb color.

hsl

string (optional), Valid hsl color.

cmyk

string (optional) Valid cmyk color.

count

integer (optional). Number of colors to return. Default: 5 Example: 6.

mode

string (optional). Define mode by which to generate the scheme from the seed color. Example: "analogic".

format

string (optional). Return results as JSON, SVG or HTML page. Example: "html".

w

integer (optional). Height of resulting image, only applicable on SVG format. Example: 350.

named

boolean (optional) Whether to print the color names on resulting image, only applicable on SVG format. Example: FALSE.

return_name_only

boolean. if FALSE returns color name only

Value

A character vector.

Examples

if (FALSE) {
get_color_thecolorapi_scheme(hex = "0047AB", count = 3, mode = "complement")
}